requirements.txt
fastmcp==2.3.4
pyodbc==5.2.0
openai==1.78.1
semantic-kernel==1.34.0
python-dotenv
